refactor(MP-34): Refactor components connected with table.
Change name from picture to image in other places accordingly.
