+++ src/components/table/Body.tsx
@@ -1,8 +1,8 @@
 import { Tbody, Td, Tr } from '@chakra-ui/react';
 import { Cell, Row, TableBodyProps } from 'react-table';
-import { ColumnInstanceDefinition } from 'model/table';
+import { ColumnInstanceDefinition } from 'model/table/table-definitions';
 import ActionCell from './ActionCell';
-import PictureCell from './PictureCell';
+import ImageCell from './ImageCell';
 
 type BodyProps<T extends object> = {
   bodyProps: TableBodyProps;
@@ -16,19 +16,16 @@ function Body<T extends object>(props: BodyProps<T>) {
       {props.rows.map((row) => {
         props.prepareRow(row);
         return (
-          <Tr borderBottom="1px" borderColor="gray.200" {...row.getRowProps()}
-              key={row.id}>
+          <Tr borderBottom="1px" borderColor="gray.200" {...row.getRowProps()} key={row.id}>
             {row.cells.map((cell: Cell<T>) => {
               // eslint-disable-next-line
               const column: ColumnInstanceDefinition<T> = cell.column;
               return (
-                <Td alignSelf="center" border="none" {...cell.getCellProps()}
-                    key={cell.column.id}>
-                  {column.isPhoto ? (
-                    <PictureCell src={cell.value} width="100px"/>
+                <Td alignSelf="center" border="none" {...cell.getCellProps()} key={cell.column.id}>
+                  {column.isImageColumn ? (
+                    <ImageCell src={cell.value} width="100px" />
                   ) : column.actions ? (
-                    <ActionCell name="Akcje" actions={column.actions}
-                                row={row}/>
+                    <ActionCell name="Akcje" actions={column.actions} row={row} />
                   ) : (
                     cell.render('Cell')
                   )}
