+++ src/components/table/Content.tsx
@@ -3,7 +3,7 @@ import { Table } from '@chakra-ui/react';
 import styled from 'styled-components';
 import Header from './Header';
 import Body from './Body';
-import { ColumnDefinition } from '../../model/table';
+import { ColumnDefinition } from '../../model/table/table-definitions';
 
 const Styles = styled.div`
   display: block;
@@ -27,13 +27,7 @@ export interface ContentProps<T extends object> {
 }
 
 function Content<T extends object>(props: ContentProps<T>) {
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    rows,
-    prepareRow
-  }: TableInstance<T> = useTable(
+  const { getTableProps, getTableBodyProps, headerGroups, rows, prepareRow }: TableInstance<T> = useTable(
     {
       columns: props.columns,
       data: props.data,
@@ -44,9 +38,8 @@ function Content<T extends object>(props: ContentProps<T>) {
   return (
     <Styles>
       <Table variant="simple" {...getTableProps()}>
-        <Header headerGroups={headerGroups}/>
-        <Body bodyProps={getTableBodyProps()} prepareRow={prepareRow}
-              rows={rows}/>
+        <Header headerGroups={headerGroups} />
+        <Body bodyProps={getTableBodyProps()} prepareRow={prepareRow} rows={rows} />
       </Table>
     </Styles>
   );
